refactor: rename mindmap_orgmode to orgmode
