+++ tests/test_mindmap_orgmode_edge_cases.py
@@ -2,7 +2,7 @@ import unittest
 import xml.etree.ElementTree as xml
 from datetime import datetime, date
 
-from mindmap_orgmode import Formatter
+from orgmode import Formatter
 
 
 class TestMindmapOrgmodeEdgeCases(unittest.TestCase):
